Add SharedSources
