+++ SharedSources/Vifra.swift
@@ -19,42 +19,22 @@ public enum Vifra {
     case strong
     /// A delay between haptics (after each successful haptic a min delay of 8500 usec is enforced).
     case delay(usec: UInt32)
-    
+}
+
+public protocol Vifrable {
     /// Blocking call to play a single haptic.
     ///
     /// - Parameter type: The type of haptic to play.
-    public static func feedback(_ type: Vifra) {
-        feedback([type])
-    }
+    static func feedback(_ type: Vifra)
     
     /// Blocking call to play multiple haptics.
     ///
     /// - Parameter types: The types of haptic to play.
-    public static func feedback(_ types: [Vifra]) {
-        Actuator.setup()
-        
-        for type in types {
-            guard let actuationID = type.actuationID else { continue }
-            
-            let params = Actuator.Parameters(arg1: 0, arg2: 1.0, arg3: 1.5)
-            Actuator.actuate(actuationID, parameters: params)
-            usleep(8500)
-        }
-        
-        Actuator.teardown()
-    }
-    
-    private var actuationID: Int32? {
-        switch self {
-        case .weak:
-            return 3
-        case .medium:
-            return 4
-        case .strong:
-            return 6
-        case let .delay(usec):
-            usleep(usec)
-            return nil
-        }
+    static func feedback(_: [Vifra])
+}
+
+public extension Vifrable {
+    public static func feedback(_ type: Vifra) {
+        feedback([type])
     }
 }
