+++ docs/Enums/Vifra.html
@@ -14,7 +14,7 @@
     <a title="Vifra Enumeration Reference"></a>
     <header>
       <div class="content-wrapper">
-        <p><a href="../index.html">Vifra Docs</a> (100% documented)</p>
+        <p><a href="../index.html">Vifra Docs</a> (75% documented)</p>
         <p class="header-right"><a href="https://github.com/matis-schotte/Vifra"><img src="../img/gh.png"/>View on GitHub</a></p>
         <p class="header-right"><a href="dash-feed://https%3A%2F%2Fmatis%2Dschotte%2Egithub%2Eio%2FVifra%2FVifra%2Exml"><img src="../img/dash.png"/>Install in Dash</a></p>
       </div>
@@ -34,6 +34,14 @@
             <ul class="nav-group-tasks">
               <li class="nav-group-task">
                 <a href="../Enums/Vifra.html">Vifra</a>
+              </li>
+            </ul>
+          </li>
+          <li class="nav-group-name">
+            <a href="../Protocols.html">Protocols</a>
+            <ul class="nav-group-tasks">
+              <li class="nav-group-task">
+                <a href="../Protocols/Vifrable.html">Vifrable</a>
               </li>
             </ul>
           </li>
@@ -183,52 +191,6 @@ to the device actuator to provide haptic feedback through the taptic engine.</p>
                 <li class="item">
                   <div>
                     <code>
-                    <a name="/s:5VifraAAO8feedbackyABFZ"></a>
-                    <a name="//apple_ref/swift/Method/feedback(_:)" class="dashAnchor"></a>
-                    <a class="token" href="#/s:5VifraAAO8feedbackyABFZ">feedback(_:)</a>
-                    </code>
-                  </div>
-                  <div class="height-container">
-                    <div class="pointer-container"></div>
-                    <section class="section">
-                      <div class="pointer"></div>
-                      <div class="abstract">
-                        <p>Blocking call to play a single haptic.</p>
-
-                      </div>
-                      <div class="declaration">
-                        <h4>Declaration</h4>
-                        <div class="language">
-                          <p class="aside-title">Swift</p>
-                          <pre class="highlight swift"><code><span class="kd">public</span> <span class="kd">static</span> <span class="kd">func</span> <span class="nf">feedback</span><span class="p">(</span><span class="n">_</span> <span class="nv">type</span><span class="p">:</span> <span class="kt">Vifra</span><span class="p">)</span></code></pre>
-
-                        </div>
-                      </div>
-                      <div>
-                        <h4>Parameters</h4>
-                        <table class="graybox">
-                          <tbody>
-                            <tr>
-                              <td>
-                                <code>
-                                <em>type</em>
-                                </code>
-                              </td>
-                              <td>
-                                <div>
-                                  <p>The type of haptic to play.</p>
-                                </div>
-                              </td>
-                            </tr>
-                          </tbody>
-                        </table>
-                      </div>
-                    </section>
-                  </div>
-                </li>
-                <li class="item">
-                  <div>
-                    <code>
                     <a name="/s:5VifraAAO8feedbackySayABGFZ"></a>
                     <a name="//apple_ref/swift/Method/feedback(_:)" class="dashAnchor"></a>
                     <a class="token" href="#/s:5VifraAAO8feedbackySayABGFZ">feedback(_:)</a>
@@ -277,7 +239,7 @@ to the device actuator to provide haptic feedback through the taptic engine.</p>
           </section>
         </section>
         <section id="footer">
-          <p>&copy; 2018 <a class="link" href="https://ungeord.net" target="_blank" rel="external">Matis Schotte</a>. All rights reserved. (Last updated: 2018-03-31)</p>
+          <p>&copy; 2018 <a class="link" href="https://ungeord.net" target="_blank" rel="external">Matis Schotte</a>. All rights reserved. (Last updated: 2018-04-01)</p>
           <p>Generated by <a class="link" href="https://github.com/realm/jazzy" target="_blank" rel="external">jazzy ♪♫ v0.9.1</a>, a <a class="link" href="http://realm.io" target="_blank" rel="external">Realm</a> project.</p>
         </section>
       </article>
