+++ docs/docsets/Vifra.docset/Contents/Resources/Documents/Protocols.html
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html lang="en">
   <head>
-    <title>Enumerations  Reference</title>
+    <title>Protocols  Reference</title>
     <link rel="stylesheet" type="text/css" href="css/jazzy.css" />
     <link rel="stylesheet" type="text/css" href="css/highlight.css" />
     <meta charset='utf-8'>
@@ -10,7 +10,7 @@
     
   </head>
   <body>
-    <a title="Enumerations  Reference"></a>
+    <a title="Protocols  Reference"></a>
     <header>
       <div class="content-wrapper">
         <p><a href="index.html">Vifra Docs</a> (75% documented)</p>
@@ -22,7 +22,7 @@
       <p id="breadcrumbs">
         <a href="index.html">Vifra Reference</a>
         <img id="carat" src="img/carat.png" />
-        Enumerations  Reference
+        Protocols  Reference
       </p>
     </div>
     <div class="content-wrapper">
@@ -49,8 +49,8 @@
       <article class="main-content">
         <section>
           <section class="section">
-            <h1>Enumerations</h1>
-            <p>The following enumerations are available globally.</p>
+            <h1>Protocols</h1>
+            <p>The following protocols are available globally.</p>
 
           </section>
           <section class="section task-group-section">
@@ -59,9 +59,9 @@
                 <li class="item">
                   <div>
                     <code>
-                    <a name="/s:5VifraAAO"></a>
-                    <a name="//apple_ref/swift/Enum/Vifra" class="dashAnchor"></a>
-                    <a class="token" href="#/s:5VifraAAO">Vifra</a>
+                    <a name="/s:5Vifra8VifrableP"></a>
+                    <a name="//apple_ref/swift/Protocol/Vifrable" class="dashAnchor"></a>
+                    <a class="token" href="#/s:5Vifra8VifrableP">Vifrable</a>
                     </code>
                   </div>
                   <div class="height-container">
@@ -69,16 +69,15 @@
                     <section class="section">
                       <div class="pointer"></div>
                       <div class="abstract">
-                        <p>Vifra is a framework for macOS, iOS and WatchKit that provides simplified access
-to the device actuator to provide haptic feedback through the taptic engine.</p>
+                        <p>Undocumented</p>
 
-                        <a href="Enums/Vifra.html" class="slightly-smaller">See more</a>
+                        <a href="Protocols/Vifrable.html" class="slightly-smaller">See more</a>
                       </div>
                       <div class="declaration">
                         <h4>Declaration</h4>
                         <div class="language">
                           <p class="aside-title">Swift</p>
-                          <pre class="highlight swift"><code><span class="kd">public</span> <span class="kd">enum</span> <span class="kt">Vifra</span></code></pre>
+                          <pre class="highlight swift"><code><span class="kd">public</span> <span class="kd">protocol</span> <span class="kt">Vifrable</span></code></pre>
 
                         </div>
                       </div>
