+++ docs/docsets/Vifra.docset/Contents/Resources/Documents/index.html
@@ -13,7 +13,7 @@
     <a title="Vifra  Reference"></a>
     <header>
       <div class="content-wrapper">
-        <p><a href="index.html">Vifra Docs</a> (100% documented)</p>
+        <p><a href="index.html">Vifra Docs</a> (75% documented)</p>
         <p class="header-right"><a href="https://github.com/matis-schotte/Vifra"><img src="img/gh.png"/>View on GitHub</a></p>
         <p class="header-right"><a href="dash-feed://https%3A%2F%2Fmatis%2Dschotte%2Egithub%2Eio%2FVifra%2FVifra%2Exml"><img src="img/dash.png"/>Install in Dash</a></p>
       </div>
@@ -33,6 +33,14 @@
             <ul class="nav-group-tasks">
               <li class="nav-group-task">
                 <a href="Enums/Vifra.html">Vifra</a>
+              </li>
+            </ul>
+          </li>
+          <li class="nav-group-name">
+            <a href="Protocols.html">Protocols</a>
+            <ul class="nav-group-tasks">
+              <li class="nav-group-task">
+                <a href="Protocols/Vifrable.html">Vifrable</a>
               </li>
             </ul>
           </li>
@@ -129,7 +137,7 @@
           </section>
         </section>
         <section id="footer">
-          <p>&copy; 2018 <a class="link" href="https://ungeord.net" target="_blank" rel="external">Matis Schotte</a>. All rights reserved. (Last updated: 2018-03-31)</p>
+          <p>&copy; 2018 <a class="link" href="https://ungeord.net" target="_blank" rel="external">Matis Schotte</a>. All rights reserved. (Last updated: 2018-04-01)</p>
           <p>Generated by <a class="link" href="https://github.com/realm/jazzy" target="_blank" rel="external">jazzy ♪♫ v0.9.1</a>, a <a class="link" href="http://realm.io" target="_blank" rel="external">Realm</a> project.</p>
         </section>
       </article>
