+++ docs/docsets/Vifra.docset/Contents/Resources/Documents/search.json
@@ -1,1 +1,1 @@
-{"Enums/Vifra.html#/s:5VifraAAO4weakA2BmF":{"name":"weak","abstract":"<p>A weak haptic.</p>","parent_name":"Vifra"},"Enums/Vifra.html#/s:5VifraAAO6mediumA2BmF":{"name":"medium","abstract":"<p>A medium haptic.</p>","parent_name":"Vifra"},"Enums/Vifra.html#/s:5VifraAAO6strongA2BmF":{"name":"strong","abstract":"<p>A strong haptic.</p>","parent_name":"Vifra"},"Enums/Vifra.html#/s:5VifraAAO5delayABs6UInt32V4usec_tcABmF":{"name":"delay","abstract":"<p>A delay between haptics (after each successful haptic a min delay of 8500 usec is enforced).</p>","parent_name":"Vifra"},"Enums/Vifra.html#/s:5VifraAAO8feedbackyABFZ":{"name":"feedback(_:)","abstract":"<p>Blocking call to play a single haptic.</p>","parent_name":"Vifra"},"Enums/Vifra.html#/s:5VifraAAO8feedbackySayABGFZ":{"name":"feedback(_:)","abstract":"<p>Blocking call to play multiple haptics.</p>","parent_name":"Vifra"},"Enums/Vifra.html":{"name":"Vifra","abstract":"<p>Vifra is a framework for macOS, iOS and WatchKit that provides simplified access"},"Enums.html":{"name":"Enumerations","abstract":"<p>The following enumerations are available globally.</p>"}}
+{"Protocols/Vifrable.html#/s:5Vifra8VifrableP8feedbackyA2AOFZ":{"name":"feedback(_:)","abstract":"<p>Blocking call to play a single haptic.</p>","parent_name":"Vifrable"},"Protocols/Vifrable.html#/s:5Vifra8VifrableP8feedbackySayA2AOGFZ":{"name":"feedback(_:)","abstract":"<p>Blocking call to play multiple haptics.</p>","parent_name":"Vifrable"},"Protocols/Vifrable.html":{"name":"Vifrable","abstract":"<p>Undocumented</p>"},"Enums/Vifra.html#/s:5VifraAAO4weakA2BmF":{"name":"weak","abstract":"<p>A weak haptic.</p>","parent_name":"Vifra"},"Enums/Vifra.html#/s:5VifraAAO6mediumA2BmF":{"name":"medium","abstract":"<p>A medium haptic.</p>","parent_name":"Vifra"},"Enums/Vifra.html#/s:5VifraAAO6strongA2BmF":{"name":"strong","abstract":"<p>A strong haptic.</p>","parent_name":"Vifra"},"Enums/Vifra.html#/s:5VifraAAO5delayABs6UInt32V4usec_tcABmF":{"name":"delay","abstract":"<p>A delay between haptics (after each successful haptic a min delay of 8500 usec is enforced).</p>","parent_name":"Vifra"},"Enums/Vifra.html#/s:5VifraAAO8feedbackySayABGFZ":{"name":"feedback(_:)","abstract":"<p>Blocking call to play multiple haptics.</p>","parent_name":"Vifra"},"Enums/Vifra.html":{"name":"Vifra","abstract":"<p>Vifra is a framework for macOS, iOS and WatchKit that provides simplified access"},"Enums.html":{"name":"Enumerations","abstract":"<p>The following enumerations are available globally.</p>"},"Protocols.html":{"name":"Protocols","abstract":"<p>The following protocols are available globally.</p>"}}
